tidy: Move scrapper under 'tools'
