+++ scripts/scrapper/intercom/metadata/main.go
@@ -5,8 +5,8 @@ import (
 	"strings"
 
 	"github.com/PuerkitoBio/goquery"
-	"github.com/amp-labs/connectors/common/scrapper"
 	"github.com/amp-labs/connectors/intercom/metadata"
+	"github.com/amp-labs/connectors/tools/scrapper"
 	"github.com/gertd/go-pluralize"
 )
