+++ tools/scrapper/models.go
