+++ tools/scrapper/utils.go
